# betbrabo-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/bets/store": {
            "post": {
                "tags": [
                    "Bets"
                ],
                "summary": "Store a newly created bet in storage",
                "operationId": "6ae0b310824d400767783c60f2e7fdeb",
                "requestBody": {
                    "description": "JSON containing bet data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "users_id",
                                    "date_start"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bet Name"
                                    },
                                    "users_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_start": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-09-27"
                                    },
                                    "date_end": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-09-30"
                                    },
                                    "date_execution": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}